Fetch events from lyyti and update google group members and google calendar events' pariticipants
