+++ src/googleservices/groups.py
@@ -68,6 +68,26 @@ def get_group_members(group_id: str) -> list[str]:
     return emails
 
 
+def list_differences(old: list[T], new: list[T]) -> tuple[list[T], list[T]]:
+    """
+    Compares two lists and returns a tuple that contains
+    1. list elements that should be added to `old` to get `new`
+    2. list elements that should be removed from `old` to get `new`
+
+    Args:
+        old (list[T]): Old list whose differences will be returned
+        new (list[T]): New list that the differences will lead to
+
+    Returns:
+        Tuple[list[T], list[T]]: Returns a tuple of two lists: (1) list elements
+        that should be added to `old` to get `new` and (2) list elements that should
+        be removed from `old` to get `new`.
+    """
+    to_be_added = set(new) - set(old)
+    to_be_removed = set(old) - set(new)
+    return list(to_be_added), list(to_be_removed)
+
+
 def update_group_members(group_id: str, new: list[str]) -> None:
     """
     Updates group members.
