+++ src/utils.py
@@ -59,21 +59,4 @@ def extract_calendar_id(calendar_url: str) -> str:
     return calendar_id
 
 
-def list_differences(old: list[T], new: list[T]) -> tuple[list[T], list[T]]:
-    """
-    Compares two lists and returns a tuple that contains
-    1. list elements that should be added to `old` to get `new`
-    2. list elements that should be removed from `old` to get `new`
 
-    Args:
-        old (list[T]): Old list whose differences will be returned
-        new (list[T]): New list that the differences will lead to
-
-    Returns:
-        Tuple[list[T], list[T]]: Returns a tuple of two lists: (1) list elements
-        that should be added to `old` to get `new` and (2) list elements that should
-        be removed from `old` to get `new`.
-    """
-    to_be_added = set(new) - set(old)
-    to_be_removed = set(old) - set(new)
-    return list(to_be_added), list(to_be_removed)
